+++ CHANGELOG.md
@@ -20,6 +20,11 @@ To see unreleased changes, please see the [CHANGELOG on the master branch](https
 
 * SentryMiddleware: `disable_integrations` flag.
 
+### Infrastructure
+
+* ruff 0.16.18
+* mypy 2.1.0
+
 ## 0.6.0 - 2025-10-10
 
 ### Added